Feat: 상세 조회 시 ArticleDetailResponse 사용

+++ src/main/java/org/sopt/domain/article/dto/ArticleDetailResponse.java
@@ -15,7 +15,7 @@ public record ArticleDetailResponse(
         String memberName,
         LocalDateTime createdAt,
         List<CommentResponse> comments,
-        int commentCount
+        Long commentCount
 ) {
     public static ArticleDetailResponse of(Article article, List<CommentResponse> comments) {
         return new ArticleDetailResponse(
@@ -26,7 +26,7 @@ public record ArticleDetailResponse(
                 article.getMember().getName(),
                 article.getCreatedAt(),
                 comments,
-                comments.size()
+                (long) comments.size()
         );
     }
 }
